BuildSticks
It's nice to have a way to visualize the connections between patches, but the least cost path analysis is slow. This tool uses the edge list produced by the 'Create edge list (Cost distance)' tool to draw lines between each of the connected patches.

I haven't tested this on any datasets other than my own. Can you take a look, and see if it gives you the correct result?

Thanks!
- Nathan