Build Sticks
It's helpful to be able to visualize the edge list, but creating least cost paths takes a significant amount of time. This uses the edge list, in combination with the patch centroids, to draw straight lines between all connected patches, which is significantly faster. If run after the Assign Subgraphs tool, this will allow different graph components to be symbolized differently.

For both this and the Assign Subgraphs tool, I've tested on my own data, but not on others. Can you take a look, and let me know if these work for you?

Thanks!
- Nathan